+++ package.json
@@ -3,7 +3,6 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "ajv": "^6.5.1",
     "axios": "^0.18.0",
     "cors": "^2.8.4",
     "prop-types": "^15.6.2",
@@ -13,7 +12,7 @@
     "react-dom": "^16.6.0",
     "react-materialize": "^2.3.1",
     "react-router-dom": "^4.2.2",
-    "react-scripts": "1.1.4",
+    "react-scripts": "^2.0.5",
     "react-transition-group": "^2.5.0"
   },
   "scripts": {
@@ -21,5 +20,11 @@
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject"
-  }
+  },
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not ie <= 11",
+    "not op_mini all"
+  ]
 }
