+++ src/Profile.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import AddDog from './addDog';
-import axios from 'axios';
+// import axios from 'axios';
 import Search from './Search';
 // import Founduser from './foundUser';
 import YourDogs from './YourDogs';
